cleaned_requirements.txt
# Core Application Dependencies
accelerate==1.6.0
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
dnspython==2.7.0
filelock==3.18.0
fsspec==2025.3.2
gradio==5.33.1
gradio_client==1.10.3
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
kombu==5.5.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
psutil==7.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
ruff==0.11.11
safehttpx==0.1.6
safetensors==0.5.3
semantic-version==2.10.0
setuptools==80.4.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1

# Environment and Configuration
python-dotenv==1.1.0

# Web Framework and API
fastapi==0.115.12
starlette==0.46.2
Flask==3.1.1
Werkzeug==3.1.3
ffmpy==0.5.0

# AI/ML Models and Processing (Updating for Hunyuan 2.1)
diffusers>=0.19.0
openai==1.82.0

# Image Processing
matplotlib>=3.3.0

# Database Support
pymongo==4.13.0

# Task Queue and Caching (CRITICAL for fixing Redis errors)
celery[redis]>=5.2.0
redis>=4.0.0
kombu[redis]>=5.2.0

# AWS Integration (Optional)
boto3>=1.26.0
botocore>=1.29.0

# 3D Processing and Geometry
trimesh>=3.15.0

# Natural Language Processing
nltk==3.9.1

# Development and Testing (Optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Utilities
pathlib2>=2.3.7
scikit-image>=0.20.0
scipy>=1.10.0
opencv-python>=4.8.0
